### 2019-11-19 #1902 @whedon
[PRE REVIEW]: OrNet - a Python Toolkit to Model the Diffuse Structure of Organelles as Social Networks
**Submitting author:** @Marcdh3 (<a href="http://orcid.org/0000-0002-9380-3181">Marcus Hill</a>)
**Repository:** <a href="https://github.com/quinngroup/ornet-JOSS" target ="_blank">https://github.com/quinngroup/ornet-JOSS</a>
**Version:** v0.1
