### 2019-11-19 #1902 @kthyng
Hi @marcdh3! Can you look through this list and suggest some appropriate reviewers for your submission? It helps to have about 5 suggestions if possible.

Please fix the citations in your paper; you can find more details here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
