### 2019-11-20 #1902 @Marcdh3
Hi @kthyng! Thank you for the quick response. Here are my reviewer suggestions:

- serine
- sonal-raj
- vc1492a
- camillescott
- nidhog

Also, I will be working on the citations now, and I'll update you once they are fixed.
