+++ paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-Mitochondria can offer key insights into the health of a human cell. These organelles are strung out across a cell in diffuse, fluctuating webs; healthy structures can be altered by disease, cellular invaders, and other malfunctioning organelles. Fluorescent tagging has shed light on mitochondria’s spatial orientation, but practical observations of structural changes are limited. Biomedical and clinical researchers cannot take advantage of large datasets (i.e. videos of mitochondria in various environments), without automated methods to quantify distribution changes over time. Fitting a model to individual video frames and comparing how the scene evolves has shown some success in measuring spatiotemporal changes (Ruan, 2019). 
+Mitochondria can offer key insights into the health of a human cell. These organelles are strung out across a cell in diffuse, fluctuating webs; healthy structures can be altered by disease, cellular invaders, and other malfunctioning organelles. Fluorescent tagging has shed light on mitochondria’s spatial orientation, but practical observations of structural changes are limited. Biomedical and clinical researchers cannot take advantage of large datasets (i.e. videos of mitochondria in various environments), without automated methods to quantify distribution changes over time. Fitting a model to individual video frames and comparing how the scene evolves has shown some success in measuring spatiotemporal changes [@Ruan:2019]. 
 
 Our project builds on this idea [@Durden:2019], as we seek to convert physical mitochondrial networks into graph networks, then characterize distinct phenotypes. This tool kit is a prototype for future analysis kits. Specifically, it focuses on characterizing two known mitochondrial conditions: drastic fragmentation and excessive fusion. This may assist researchers in larger clinical studies, but expanding this project might eventually yield tools to pick up on structural changes that elude the human eye.
 
