### 2019-11-20 #1902 @Marcdh3
@kthyng I believe we fixed the citations. Please let us know if there is anything else needing to be fixed.
