### 2019-11-21 #1902 @akeshavan
:wave: @camillescott and @nidhog: would you be willing to review this paper for JOSS?
