### 2019-11-23 #1902 @arfon
> @whedon add @akeshavan as editor

@kthyng it's `assign` not `add` for the editor. Sorry for any confusion caused.
