### 2019-11-25 7e61184
Modified the pipeline to use Hellinger distance
