+++ src/measure.py
@@ -53,3 +53,29 @@ def multivariate_js(m1, s1, m2, s2):
     pq = multivariate_kl(m1, s1, m2, s2)
     qp = multivariate_kl(m2, s2, m1, s1)
     return 0.5 * (pq + qp)
+
+def multivariate_hellinger(u1, cov1, u2,cov2, gamma = 0.00125):
+	"""
+	Hellinger distance measure for high-dimensional gaussians. For
+	use with analytic parameters, rather than evaluating empirical distributions.
+
+	For the sake of notation: N(m1, s1) ~ p, and N(m2, s2) ~ q
+
+	Parameters
+	----------
+	u1, u2 : array, shape (d,)
+	    Means of the two distributions.
+	cov1, cov2 : array, shape (d, d)
+            Covariance matrices of the two distributions.
+
+	Returns
+	-------
+	hellinger : float
+            Hellinger distance.
+	"""
+
+	mcov = 0.5*cov1+0.5*cov2
+	dets = np.sqrt(np.sqrt(np.linalg.det(cov1))*np.sqrt(np.linalg.det(cov2))/np.linalg.det(mcov))
+	mahala = (u1-u2).dot(np.linalg.pinv(mcov)).dot(u1-u2)
+	h = np.exp(-gamma * mahala)*dets
+	return 1 - np.sqrt(1-h)
