### 2019-11-27 #1902 @akeshavan
👋 @serine, @sonal-raj, and @vc1492a -- would you be willing to review this paper for JOSS?
