### 2019-11-28 #1902 @serine
This looks interesting. I should be able to take a look at it next week, if I can get a hold of raw/example data. Image analysis isn't my field so I wouldn't know where to source the data from.

One super minor thing, but it feels rather weird to call the repository (repo) `ornet-JOSS`. I'm assuming **JOSS** suffix in the name referres to the fact that the tool will be submitted to JOSS for review? If this is the case I think it would be much better to drop **JOSS** from the repo name, while this is still early days for the repo.

Usually repo's name hints / describes what the content of the repo is and I'd say for a lot of the source code the repo name usually the name of the tool/library. JOSS `paper.md` is a very small part of the repo.

Just think about establishing a name for your tool and advertising it to other people.

Cheers
