### 2019-12-06 #1902 @vc1492a
@akeshavan thanks for reaching out about the opportunity to review this work. I'll be taking a long vacation in about a week's time and will not be returning until close to the end of January. If that's all right with you and the authors, I'm happy to review this work. However I'll be largely unavailable during that time due to travel.

The work is interesting although it's not typically a field I'm involved in, so my contributions for review in regards to it's utility in the field will be very limited. I can however review for code organization, completeness, documentation suitability, etc. In line with @serine, I'd need to get a hold of some example data in order to review this work effectively - ideally more than one sample in which to test.

Hope this helps!
