### 2019-12-09 a13be98
Added the median normalization code and tests
