+++ src/OrNetSegmentation.py
@@ -1,135 +1,0 @@
-import numpy as np
-import cv2
-from matplotlib import pyplot as plt
-import imageio
-import numpy as np
-import random
-
-def main():
-	# First we read the manual segmentation and copy it to 'im'
-	filename = 'DsRed2-HeLa_2_21_LLO'
-	foldername = 'LLO'
-	im = imageio.imread('First Frames Segments/'+foldername+'/'+filename+'.vtk')#Directory will be machine dependant
-	plt.imshow(im)
-	plt.show()
-
-	#=======================================================================#
-	#																		#
-	#=======================================================================#
-
-	video_file ='Videos/'+foldername+'/'+filename+' (Converted).mov' #Directory is machine dependant
-	vf = cv2.VideoCapture(video_file)	#OpenCV videocapture object
-
-	# obtaining the frame dimensions
-	X = int(vf.get(3)) #vf.get returns specific properties, 3 is width 4 is height
-	Y = int(vf.get(4))
-
-	print (X, Y)
-
-	# Setting the variables for saving the out put video like File name , encoding method , fps and ...
-	fourcc = cv2.VideoWriter_fourcc(*'MJPG') 
-	out = cv2.VideoWriter('test1.mov',fourcc, 30.0, (X, Y)) #OpenCV VideoWriter Object
-	
-
-
-	frameNum = 0
-	number_of_segments = len(np.unique(im))-1 #defines number of segs from vtk
-	print(number_of_segments)
-	outs = list()
-	for i in range(number_of_segments):
-		outs.append(cv2.VideoWriter(filename+'_Cell'+str(i)+'.mov',fourcc,30.0, (X, Y)))
-
-	ims = list () #holds masked images
-	masks = list() #holds masks
-	opening = list() #holds images after opened
-	colors = list() #the colors to make the contours in the output video?
-	contours = list() #holds contours
-	dilates = list() #holds dilated shapes
-	outframes = list()
-	first = True #determines first frame
-	kernel = np.ones((17,17), np.uint8) #kernel for opening
-	kernel2 = np.ones((3,3),np.uint8) #kernel for dilation
-	font = cv2.FONT_HERSHEY_SIMPLEX #font for frame count
-
-	for cols in range (number_of_segments): #creates random colors to use for the outlines
-		colors.append((random.randint(1,255), random.randint(0,255), random.randint(1,255)))
-
-
-	while( vf.isOpened() ):
-		#for i in range(3):#skips to every third frame
-		frameNum += 1 #advance through frames
-		ret, frame = vf.read()
-	
-		segsGrow = np.ones(number_of_segments,dtype = bool)
-		for i in range (number_of_segments): #adds a copy of the current frame for each segment
-			ims.append(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY))
-			outframes.append(np.copy(frame))
-		for i in range ( number_of_segments ): #separates each mask from the vtk and lists them
-			masks.append( im != i + 1 )
-		for i in range (number_of_segments): #blacks out all that isn't in the initial mask for each segment
-			if(first):
-				ims[i][masks[i]] = 0
-			else:
-				ims[i][dilates[i]!=255] = 0 #uses previous dilation for next frame
-			
-		del dilates[:]
-		for i in range(number_of_segments): #thresholds the image with OTSU's optimal threshold value
-			#ims[i] = cv2.GaussianBlur(ims[i],(5,5),0) 
[…]
