### 2019-12-14 #1902 @Marcdh3
I've uploaded some videos that we've been using in the "samples" directory. Inside of that directory are three sub-directories labeled "control", "llo", and "mdivi". Each sub-directory contains a video (.mov) and an initial segmentation mask (.vtk). The sub-directories correspond to video data of HeLa cells that were either in the control group, exposed to pre-forming toxin listeri-olysin O (LLO), or exposed to mitochondria division inhibitor-1 (mdivi-1). These videos should do a good job illustrating the morphology changes of mitochondria under different conditions.
