### 2020-01-03 40ce3bc
Using RGB instead of BGR functions. Sped up median normalization using numpy indexing instead of loops. Renamed function parameters in Pipeline.py for clarity.
