### 2020-01-07 5fc4b3f
formatting and cleanup
