+++ src/Pipeline.py
@@ -4,7 +4,7 @@ The pipeline consists of cell segmentation, graph vertex construction
 via Gaussian mixture model means, edge construction via divergence 
 functions, and eigen-decomposition of the matrix representation.
 '''
-#Author: Marcus Hill
+# Author: Marcus Hill
 
 import os
 import re
@@ -23,333 +23,356 @@ from ornet.affinityfunc import get_all_aff_tables
 from ornet.extract_helper import generate_singles
 from ornet.median_normalization import median_normalize as normalize
 
+
 def constrain_vid(vid_path, full_path, frame_num):
-	'''
-	Constrains the input video to specified number of frames, and write the 
-	result to an output video. If the video contains less frames than 
-	frame_num, then then all frames of the video are returned.
+    '''
+    Constrains the input video to specified number of frames, and write the
+    result to an output video. If the video contains less frames than
+    frame_num, then then all frames of the video are returned.
 
-	Parameters
-	----------
-	vid_path: String
-		Path to the input video.
-	full_path: String
-		Path to the output video.
-	frame_num: int
-		Maximum number of frames to extract from the video.
+    Parameters
+    ----------
+    vid_path: String
+        Path to the input video.
+    full_path: String
+        Path to the output video.
+    frame_num: int
+        Maximum number of frames to extract from the video.
 
-	Returns
-	----------
-	NoneType object
-	'''
-	reader = imageio.get_reader(vid_path)
-	fps = reader.get_meta_data()['fps']
-	size = reader.get_meta_data()['size']
-	writer = cv2.VideoWriter(full_path, 
-			cv2.VideoWriter_fourcc('M','J','P','G'), 
-			fps, size)
+    Returns
+    ----------
+    NoneType object
+    '''
+    reader = imageio.get_reader(vid_path)
+    fps = reader.get_meta_data()['fps']
+    size = reader.get_meta_data()['size']
+    writer = cv2.VideoWriter(full_path,
+                             cv2.VideoWriter_fourcc('M', 'J', 'P', 'G'),
+                             fps, size)
 
-	i = 0
-	for frame in reader:
-		if i == frame_num:
-			break
-		else:
-			writer.write(frame)
-			i += 1
+    i = 0
+    for frame in reader:
+        if i == frame_num:
+            break
+        else:
+            writer.write(frame)
+            i += 1
 
-	reader.close()
-	writer.release()
+    reader.close()
+    writer.release()
+
 
 def cell_segmentation(vid_name, vid_path, masks_path, out_path):
-	'''
-	Generates segmentation masks for every frame in the video, and saves 
-	the output at the specified output path.
+    '''
+    Generates segmentation masks for every frame in the video, and saves
+    the output at the specified output path.
 
-	Parameters
-	----------
-	vid_path: String
-		Path to input video.
-	masks_path: String
-		Path to initial segmentation mask.
-	out_path: String
-		Path to output directory.
+    Parameters
+    ----------
+    vid_path: String
+        Path to input video.
+    masks_path: String
+        Path to initial segmentation mask.
+    out_path: String
+        Path to output directory.
 
-	Returns
-	----------
-	NoneType object
-	'''
-	masks = extract_cells(vid_path, masks_path, show_video=False)
-	np.save(os.path.join(out_path, vid_name + 'MASKS.npy'), masks)
+    Returns
+    ----------
+    NoneType object
+    '''
+    masks = extract_cells(vid_path, masks_path, show_video=False)
+    np.save(os.path.join(out_path, vid_name + 'MASKS.npy'), masks)
+
 
[…]
