### 2020-01-07 #1983 @vc1492a
**Overview**: This was an interesting piece of software to review as its use case it outside of my typical area, but makes use of many Python software packages I use in my work and side projects. I appreciate the authors commitment to providing docstring throughout the code base which allowed me to dive in more deeply and provide some suggestions for further improvement.

Below are my comments on particular sections, of which most (if not all) are relatively minor. When the authors strictly meet the requirements laid out in the review checklist, I have marked these items as complete. For those where the requirement is not strictly met, I have left the checklist entry empty. Either way, I have provided some comments below with actions that should be taken (in the case of unchecked entries) and actions that could be taken to improve the work further but that aren't strictly required for submission to JOSS (in the case of checked entries).

**Repository URL**: While the software is available at the repository URL, the URL listed in this issue as part of the review textually does not match the destination (https://github.com/quinngroup/ornet-JOSS vs https://github.com/quinngroup/ornet). This should be noted and corrected prior to publication to JOSS by the editor.

**Installation and Installation Instructions**: While the installation proceeds without issue according to the instructions provided in `readme.md` and has clearly listed dependencies in both `readme.md` and `setup.py`, I did not see a `requirements.txt` file that is typical with Python software that can be used to install the necessary dependencies in the following way:

```python
pip install -r requirements.txt
```

It is recommended that the authors include a `requirements.txt` file in their repository, and I have provided a version of this file [as part of this pull request](https://github.com/quinngroup/ornet/pull/11) (which can be generated using the command `pip freeze > requirements.txt`). The authors should check this file and make any adjustments as necessary.

In checking the agreement between the dependencies in `readme.md` and `setup.py`, I noticed that the `opencv-python` dependency in `setup.py` is commented and not included as part of a fresh install. If it's possible to install using `setup.py`, then the line should be uncommented such that the installation can proceed in that way. I noticed that it is a required package for running the tests, so I have fixed the issue in `setup.py` [as part of this pull request](https://github.com/quinngroup/ornet/pull/11).

Lastly, it would be beneficial for the authors to consider using [semantic versioning](https://semver.org/) if not doing so already, consider releasing their software on [PyPi](https://pypi.org/) for easy installation using a command like `pip install ornet`, and tagging releases on git and uploading those releases to Github. It may be beneficial to tag the releases that are uploaded to Github and/or PyPi versioned using semantic versioning such that users can download the software directly from Github via a release as opposed to having to clone or pull the repository.

**Functionality, Functionality Documentation, and Example Usage**: Running the samples took quite a bit of time on the machine in which I was using to review the software (greater than 11 hours, almost all in video processing), which is energy-efficient and not designed for large workloads (e.g. 1.1 Ghz Intel Core m3). It may be beneficial for the authors to note the expected duration of running OrNet on the sample videos and masks or provide suggested hardware requirements.

The authors did include sample (example) files for understanding how to use the functionality of OrNet and included the necessary commands needed to use the software (e.g. functionality documentation). However, it would be good to explicitly note in the documentation (e.g. in `readme.md`) that there are examples in the `samples` directory in which to try out the software's functionality. This is implied by the existence of the samples directory but it isn't explicit in the documentation.

**Automated Tests**: After ensuring `opencv-python` was installed, I was able to run the commands that execute the tests according to the instructions provided by the authors. However, one of the 7 tests fail when running on my machine (macOS version 10.14.6 with Python version 3.7.2), specifically line 58 in `ornet_tests.py`. The authors should address this prior to publication in JOSS. The results show 7 tests with 1 failure, but the readme states that there are 6 tests - this discrepancy ought to be addressed as well.

If the authors are looking to have others contribute to the project more easily in the future, they may want to consider setting up a continuous integration pipeline that will automatically run tests and report the test results (e.g. using Travis CI) and unit test coverage (e.g. using Coveralls) with any pushes to `master` or `dev` and for any pull requests opened by open-source contributors. This pipeline would ease the burden on making sure new additions and/or changes to software work as intended, and would also allow maintainers to formally test against every version of Python 3 that should be supported. This would allow the authors to state "this software works in Python versions 3.7 - 3.8" as opposed to simply stating something along the lines of "works in Python >=3.7" in the readme.

**State of the Field**: The paper does a great job referencing how the software has been built on top of already-existing and well-utilized Python libraries such as [scikit-learn](https://scikit-learn.org/) or [matplotlib](https://matplotlib.org/). In addition, the paper makes reference to a C software called ITK-SNAP to generate masks for analysis by OrNet. However, the software paper does not explicitly state whether OrNet represents a new capability in the field or whether it ought to be compared with any existing software packages. The authors may want to consider being more explicit in the wording of the paper in this area - if its a new capability, explicitly make the claim. Otherwise, compare OrNet to any existing software packages that allow researchers to perform a similar task. This perhaps could be as simple as stating that previously researchers were required to build their own scripts for this task manually and on their own.

**References**: The two references provided are appropriately cited and relevant to this work and the software paper that is included in this submission, and are in the proper citation syntax. However, I feel that additional references would make the supplied software paper much stronger, especially for readers / reviewers that are not in the authors' field of study. For example, the first paragraph states "Fluorescent tagging has shed light on mitochondria's spatial orientation...", but does not provide a reference that indicates the first occurrence of this or any seminal works. Additionally, while it may seem odd at first, the Python language was originally presented at a conference and can thus be referenced (I have supplied this reference in `paper.bib` as part of [this pull request](https://github.com/quinngroup/ornet/pull/11), but `paper.md` should be amended by the authors). Similarly, [scikit-learn should be referenced](https://scikit-learn.org/stable/about.html) as well as [software in the scipy ecosystem](https://www.scipy.org/citing.html) (I have supplied the scikit-learn reference in `paper.bib` as part of [this pull request](https://github.com/quinngroup/ornet/pull/11), but `paper.md` should be amended by the authors for this as well as scipy prior to publication).

**Community Guidelines**: The authors have a section dedicated in the readme for users of the software to request changes to the software, report any issues during use, and ask any questions (to be performed by opening issues within the Github repository). However, the section does not outline specifically _how_ users and/or developers may contribute to the software for future use. It may benefit the authors / developers to include more specific contribution instructions, such as asking users to first fork the repository and then create a new branch with a name that describes the issue or new feature before submitting a pull request. This may result in some additional organization and encourage others to contribute to the software.

**Other**: Running the samples takes a very long time on hardware that isn't optimized for video processing, like the hardware I used in my review (just over 11 hours to normalize the input video). There are perhaps a few areas in the code base that could take advantage of multi-processing or parallelism to speed up computation, which would be helpful in a practical sense, and I highly suggest introducing progress bars or other console output that will help inform the user of how much progress has been made when using OrNet. I've performed some code formatting and cleanup that is consistent with [PEP8 formatting](https://www.python.org/dev/peps/pep-0008/) as part of the aforementioned pull request. Other small tweaks which could be made: the gamma parameter is not included in the docstring for the function `multivariate_hellinger` in `measure.py` (as well as other parameters in other files such as `run_gmm.py` that are not in dosctring), using typing for functions which was introduced in Python 3.5, and removing unused code in `extract_cells.py`.
