### 2020-01-09 #1983 @Marcdh3
@vc1492a Thank you so much for the detailed review of our software! We really appreciate the effort you made into making that pull request. As of now, that request has been merged, and we are working on incorporating all of the other changes you pointed out. Also, we have improved the median normalization code so it no longer takes hours but minutes, so if you ever get free time and want to play around with it again it should be a much better experience. We will definitely update everyone once all of those changes have been made.
