### 2020-01-09 2c778c4
Updated the number of tests in README.md
