### 2020-01-10 17ac84d
Added travis yaml. Added coverage to requirements.txt
