+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+python:
+  - "3.6"
+  - "3.7"
+  - "3.8"
+install:
+  - pip install -U pip
+  - pip install -r requirements.txt   
+  - pip install .
+script:
+  - coverage run tests/ornet_tests.py
