### 2020-01-10 ddaa225
Removed ornet from requirements.txt. Added a badge to README.md
