+++ requirements.txt
@@ -17,7 +17,6 @@ matplotlib==3.1.2
 networkx==2.4
 numpy==1.18.1
 opencv-python==4.1.2.30
-ornet==0.1
 Pillow==7.0.0
 pyparsing==2.4.6
 python-dateutil==2.8.1
