### 2020-01-10 6d3ac49
Updated supported python versions
