### 2020-01-11 8e70bec
A build failed was not encountered when a push had failures, so the travis yaml has been edited to test for this.
