+++ .travis.yml
@@ -8,4 +8,7 @@ install:
   - pip install -r requirements.txt   
   - pip install .
 script:
-  - coverage run tests/ornet_tests.py
+  - cd tests
+  - python -m unittest discover -s .
+after_success:
+  - pwd
