+++ .travis.yml
@@ -8,7 +8,6 @@ install:
   - pip install -r requirements.txt   
   - pip install .
 script:
-  - cd tests
-  - python -m unittest discover -s .
+  - python -m unittest discover -s tests
 after_success:
-  - pwd
+  - ehco $(pwd)
