+++ .travis.yml
@@ -9,5 +9,6 @@ install:
   - pip install .
 script:
   - python -m unittest discover -s tests
-after_success:
+  - coverage run -m unittest discover -s tests
+after_script:
   - ehco $(pwd)
