+++ .travis.yml
@@ -8,7 +8,4 @@ install:
   - pip install -r requirements.txt   
   - pip install .
 script:
-  - python -m unittest discover -s tests
   - coverage run -m unittest discover -s tests
-after_script:
-  - ehco $(pwd)
