### 2020-01-11 011240e
build should pass. Test what directory the system is in after the script is ran
