+++ .travis.yml
@@ -8,4 +8,8 @@ install:
   - pip install -r requirements.txt   
   - pip install .
 script:
-  - coverage run -m unittest discover -s tests
+  - cd tests
+  - coverage run --source=ornet -m unittest discover -s .
+after_success:
+  - pwd
+  - cd tests
