### 2020-01-11 b486df9
All builds successful. Cleaned up travis yaml.
