### 2020-01-14 328c086
Moved packages from to requirements to .travis.yml
