+++ requirements.txt
@@ -1,4 +1,3 @@
-coverage==5.0.2
 cycler==0.10.0
 Cython==0.29.14
 decorator==4.4.1
