+++ .travis.yml
@@ -6,6 +6,7 @@ python:
 install:
   - pip install -U pip
   - pip install -r requirements.txt   
+  - pip install coverage coveralls
   - pip install .
 script:
   - cd tests
