+++ paper.bib
@@ -1,3 +1,114 @@
+@ARTICLE{ behnel2010cython,
+    author={Behnel, S. and Bradshaw, R. and Citro, C. and Dalcin, L. and Seljebotn, D.S. and Smith, K.},
+    journal={Computing in Science Engineering},
+    title={Cython: The Best of Both Worlds},
+    year={2011},
+    month=march-april ,
+    volume={13},
+    number={2},
+    pages={31 -39},
+    keywords={Cython language;Fortran code;Python language extension;numerical loops;programming language;C language;numerical analysis;},
+    doi={10.1109/MCSE.2010.118},
+    ISSN={1521-9615},
+}
+
+@software{Clark:2016,
+    author={wiredfool, Alex Clark, Hugo, Andrew Murray, Alexander Karpinsky, Christoph Gohlke, Brian Crowell, David Schmidt, Alastair Houghton, Steve Johnson, Sandro Mani, Josh Ware, David Caro, Steve Kossouho, Eric W. Brown, ThoughtSynth, Antony Lee, Mikhail Korobov, Michał Górny, Esteban Santana Santana, Nicolas Pieuchot, Oliver Tonnhofer, Michael Brown, Benoit Pierre, Joaquín Cuenca Abela, Lars Jørgen Solberg, Felipe Reyes, Alexey Buzanov, Yifu Yu, eliempje, Fredrik Tolf}
+    title={Pillow},
+    year={2016}
+    doi={10.5281/zenodo.44297}
+}
+
+@article{Durden:2019,
+    author={Andrew Durden},
+    title={Modelling Diffuse Subcellular Protein Structures as Dynamic Social Networks},
+    year={2019},
+    eprint={1904.12960},
+    archivePrefix={arXiv},
+    primaryClass={q-bio.SC}
+}
+
+@article{Hagberg:2008,
+    author={Eric A. Hagberg, Daniel A. Schult and Pieter J. Swart},
+    title={Exploring network structure, dynamics, and function using NetworkX},
+    journal={Proceedings of the 7th Python in Science Conference (SciPy2008)},
+    pages={11--15},
+    month={Aug},
+    year={2008}
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@misc{kiwisolver2018,
+ month = {5},
+ title = {kiwisolver},
+ version = {rev 5dtKQ1Z3cpsFKT8nfQEpQhw3PNcqim},
+ year = {2018}
+}
+
+@software{Klein:2019,
[…]
