+++ paper.bib
@@ -1,4 +1,4 @@
-@ARTICLE{ behnel2010cython,
+@ARTICLE{ behnel:2010,
     author={Behnel, S. and Bradshaw, R. and Citro, C. and Dalcin, L. and Seljebotn, D.S. and Smith, K.},
     journal={Computing in Science Engineering},
     title={Cython: The Best of Both Worlds},
@@ -93,7 +93,7 @@ doi=10.5555/1406599,
   note = {[Online; accessed <today>]}
  }
 
-@article{opencv_library,
+@article{opencv_library:2000,
     author = {Bradski, G.},
     journal = {Dr. Dobb's Journal of Software Tools},
     keywords = {bibtex-import},
