### 2020-01-15 d24b27f
Updated dated paper with references to the new citations
