### 2020-01-15 4033652
Added missing curly braces, and removed non-keywords
