+++ paper.bib
@@ -1,9 +1,9 @@
-@ARTICLE{ behnel:2010,
+@ARTICLE{behnel:2010,
     author={Behnel, S. and Bradshaw, R. and Citro, C. and Dalcin, L. and Seljebotn, D.S. and Smith, K.},
     journal={Computing in Science Engineering},
     title={Cython: The Best of Both Worlds},
     year={2011},
-    month=march-april ,
+    month={march-april},
     volume={13},
     number={2},
     pages={31 -39},
@@ -49,7 +49,7 @@
   image generation across user interfaces and operating systems.},
   publisher = {IEEE COMPUTER SOC},
   doi       = {10.1109/MCSE.2007.55},
-  year      = 2007
+  year      = {2007}
 }
 
 @misc{kiwisolver2018,
@@ -80,8 +80,8 @@ title = {Getting Started with Pyparsing},
 year = {2007},
 isbn = {9780596514235},
 publisher = {O’Reilly},
-edition = {First}
-doi=10.5555/1406599,
+edition = {First},
+doi={10.5555/1406599}
 }
 
 @Misc{numpy,
@@ -89,7 +89,7 @@ doi=10.5555/1406599,
   title =     {{NumPy}: A guide to {NumPy}},
   year =      {2006--},
   howpublished = {USA: Trelgol Publishing},
-  url = "http://www.numpy.org/",
+  url = {"http://www.numpy.org/"},
   note = {[Online; accessed <today>]}
  }
 
@@ -110,15 +110,14 @@ doi=10.5555/1406599,
 }
 
 @article{Ruan:2019,
-  	Adsurl = {https://www.biorxiv.org/content/biorxiv/early/2019/01/16/522763.full.pdf},
   	Archiveprefix = {bioRxiv},
   	Author = {{Ruan}, X. and {Johnson}, G.~R. and {Johnson}, I.~B. and {Nitschke}, R. and {Boerries}, M. and {Busch}, H. and {Murphy}, R.~F.},
   	DOI = {10.1101/522763},
   	Journal = {BioRXiv},
   	Keywords = {Cell Biology},
-  	Month = jan,
+  	Month = {jan},
   	Title = {{Image-derived Models of Cell Organization Changes During Differentiation of PC12 Cells}},
-  	Year = 2019
+  	Year = {2019}
 }
 
 @software{van der Walt:2014,
@@ -158,8 +157,8 @@ doi=10.5555/1406599,
         title = "{SciPy 1.0--Fundamental Algorithms for Scientific
                   Computing in Python}",
       journal = {arXiv e-prints},
-         year = "2019",
-        month = "Jul",
+         year = {"2019"},
+        month = {"Jul"},
           eid = {arXiv:1907.10121},
         pages = {arXiv:1907.10121},
 archivePrefix = {arXiv},
