+++ paper.md
@@ -38,7 +38,7 @@ Mitochondria can offer key insights into the health of a human cell. These organ
 
 Our project builds on this idea [@Durden:2019], as we seek to convert physical mitochondrial networks into graph networks, then characterize distinct phenotypes. This tool kit is a prototype for future analysis kits. Specifically, it focuses on characterizing two known mitochondrial conditions: drastic fragmentation and excessive fusion. This may assist researchers in larger clinical studies, but expanding this project might eventually yield tools to pick up on structural changes that elude the human eye.
 
-OrNet, a python package, relies heavily on algorithms from the Scikit-Learn tool kit [@scikit-learn:2011]. This in turn is built on NumPy [@numpy], SciPy [@2019arXiv190710121V], and Matplotlib [@Hunter:2007]. Our program takes videos of human cells with fluorescently tagged mitochondria and first extracts individual cells. With that base to work from, a model is fitted to video frames by calculating a mixture of probability masses (Gaussian Mixture Model). Essentially, this model identifies the center of the densest regions of mitochondrial protein, or peaks, and finds the probability that the surrounding protein belongs to a specific local distribution. The ultimate goal is to analyze the data as a graph, where the peaks in density are used as nodes, and probability metrics form weights and edges between nodes. 
+OrNet, a python package [@Python:1995], relies heavily on algorithms from the Scikit-Learn tool kit [@scikit-learn:2011]. This in turn is built on NumPy [@Numpy:2006], SciPy [@Scipy:2019], and Matplotlib [@Hunter:2007]. Our program takes videos of human cells with fluorescently tagged mitochondria and first extracts individual cells. With that base to work from, a model is fitted to video frames by calculating a mixture of probability masses (Gaussian Mixture Model). Essentially, this model identifies the center of the densest regions of mitochondrial protein, or peaks, and finds the probability that the surrounding protein belongs to a specific local distribution. The ultimate goal is to analyze the data as a graph, where the peaks in density are used as nodes, and probability metrics form weights and edges between nodes. 
 
 The current version requires segmentation masks for each cell in a video, taken from the initial frame of each video, to seed the complete segmentation algorithm. Past masks were generated using the C based ITK-SNAP medical imaging tool [@Yoo:2002]. Running the current code requires executing the Pipeline.py script with arguments denoting the location of raw videos, masks, and desired path for output. Output takes the form of learned components from the mixture model and the calculated weights between nodes. These outputs are ripe for analysis at the researcher's discretion.
 
