+++ paper.bib
@@ -84,7 +84,7 @@ edition = {First},
 doi={10.5555/1406599}
 }
 
-@Misc{numpy,
+@Misc{Numpy:2006,
   author =    {Travis Oliphant},
   title =     {{NumPy}: A guide to {NumPy}},
   year =      {2006--},
@@ -140,7 +140,7 @@ doi={10.5555/1406599}
  year={2011}
 }
 
-@ARTICLE{2019arXiv190710121V,
+@ARTICLE{Scipy:2019,
        author = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant},
          Travis E. and {Haberland}, Matt and {Reddy}, Tyler and
          {Cournapeau}, David and {Burovski}, Evgeni and {Peterson}, Pearu
@@ -157,8 +157,8 @@ doi={10.5555/1406599}
         title = "{SciPy 1.0--Fundamental Algorithms for Scientific
                   Computing in Python}",
       journal = {arXiv e-prints},
-         year = {"2019"},
-        month = {"Jul"},
+         year = {2019},
+        month = {Jul},
           eid = {arXiv:1907.10121},
         pages = {arXiv:1907.10121},
 archivePrefix = {arXiv},
