+++ paper.bib
@@ -13,7 +13,7 @@
 }
 
 @software{Clark:2016,
-    author={wiredfool, Alex Clark, Hugo, Andrew Murray, Alexander Karpinsky, Christoph Gohlke, Brian Crowell, David Schmidt, Alastair Houghton, Steve Johnson, Sandro Mani, Josh Ware, David Caro, Steve Kossouho, Eric W. Brown, ThoughtSynth, Antony Lee, Mikhail Korobov, Michał Górny, Esteban Santana Santana, Nicolas Pieuchot, Oliver Tonnhofer, Michael Brown, Benoit Pierre, Joaquín Cuenca Abela, Lars Jørgen Solberg, Felipe Reyes, Alexey Buzanov, Yifu Yu, eliempje, Fredrik Tolf}
+    author={Alex Clark, wiredfool, Hugo, Andrew Murray, Alexander Karpinsky, Christoph Gohlke, Brian Crowell, David Schmidt, Alastair Houghton, Steve Johnson, Sandro Mani, Josh Ware, David Caro, Steve Kossouho, Eric W. Brown, ThoughtSynth, Antony Lee, Mikhail Korobov, Michał Górny, Esteban Santana Santana, Nicolas Pieuchot, Oliver Tonnhofer, Michael Brown, Benoit Pierre, Joaquín Cuenca Abela, Lars Jørgen Solberg, Felipe Reyes, Alexey Buzanov, Yifu Yu, eliempje, Fredrik Tolf}
     title={Pillow},
     year={2016}
     doi={10.5281/zenodo.44297}
@@ -91,7 +91,7 @@ doi={10.5555/1406599}
   howpublished = {USA: Trelgol Publishing},
   url = {"http://www.numpy.org/"},
   note = {[Online; accessed <today>]}
- }
+}
 
 @article{opencv_library:2000,
     author = {Bradski, G.},
