### 2020-01-15 caa197b
Updatd paper
