+++ paper.bib
@@ -13,9 +13,9 @@
 }
 
 @software{Clark:2016,
-    author={Alex Clark, wiredfool, Hugo, Andrew Murray, Alexander Karpinsky, Christoph Gohlke, Brian Crowell, David Schmidt, Alastair Houghton, Steve Johnson, Sandro Mani, Josh Ware, David Caro, Steve Kossouho, Eric W. Brown, ThoughtSynth, Antony Lee, Mikhail Korobov, Michał Górny, Esteban Santana Santana, Nicolas Pieuchot, Oliver Tonnhofer, Michael Brown, Benoit Pierre, Joaquín Cuenca Abela, Lars Jørgen Solberg, Felipe Reyes, Alexey Buzanov, Yifu Yu, eliempje, Fredrik Tolf}
+    author={Alex Clark, Andrew Murray, Alexander Karpinsky, Christoph Gohlke, Brian Crowell, David Schmidt, Alastair Houghton, Steve Johnson, Sandro Mani, Josh Ware, David Caro, Steve Kossouho, Eric W. Brown, Antony Lee, Mikhail Korobov, Michał Górny, Esteban Santana Santana, Nicolas Pieuchot, Oliver Tonnhofer, Michael Brown, Benoit Pierre, Joaquín Cuenca Abela, Lars Jørgen Solberg, Felipe Reyes, Alexey Buzanov, Yifu Yu, Fredrik Tolf, wiredfool, Hugo, eliempje, ThoughtSynth},
     title={Pillow},
-    year={2016}
+    year={2016},
     doi={10.5281/zenodo.44297}
 }
 
@@ -89,11 +89,11 @@ doi={10.5555/1406599}
   title =     {{NumPy}: A guide to {NumPy}},
   year =      {2006--},
   howpublished = {USA: Trelgol Publishing},
-  url = {"http://www.numpy.org/"},
+  url = {http://www.numpy.org/},
   note = {[Online; accessed <today>]}
 }
 
-@article{opencv_library:2000,
+@article{opencv_library,
     author = {Bradski, G.},
     journal = {Dr. Dobb's Journal of Software Tools},
     keywords = {bibtex-import},
@@ -120,7 +120,7 @@ doi={10.5555/1406599}
   	Year = {2019}
 }
 
-@software{van der Walt:2014,
+@software{scikit-image:2014,
     author={Stéfan van der Walt, Johannes L. Schönberger, Juan Nunez-Iglesias, François Boulogne, Joshua D. Warner, Neil Yager, Emmanuelle Gouillart, Tony Yu, and the scikit-image contributors},
     title={scikit-image: Image processing in Python},
     eprint={PeerJ 2:e453},
@@ -140,7 +140,7 @@ doi={10.5555/1406599}
  year={2011}
 }
 
-@ARTICLE{Scipy:2019,
+@ARTICLE{2019arXiv190710121V,
        author = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant},
          Travis E. and {Haberland}, Matt and {Reddy}, Tyler and
          {Cournapeau}, David and {Burovski}, Evgeni and {Peterson}, Pearu
@@ -157,8 +157,8 @@ doi={10.5555/1406599}
         title = "{SciPy 1.0--Fundamental Algorithms for Scientific
                   Computing in Python}",
       journal = {arXiv e-prints},
-         year = {2019},
-        month = {Jul},
+         year = "2019",
+        month = "Jul",
           eid = {arXiv:1907.10121},
         pages = {arXiv:1907.10121},
 archivePrefix = {arXiv},
@@ -174,4 +174,3 @@ archivePrefix = {arXiv},
     year={2002},
     publisher={IOS Press Amsterdam}
 }
-
