### 2020-01-15 5250a8b
Created a Community Guidelines section
