+++ README.md
@@ -6,15 +6,14 @@ Tracking *diffuse* organelles presents more challenges than just characterizing 
 
 *Mitochondria in a single cell, with red flourescent tagging* 
 
-# Dependencies
-**Required packages:** scipy, numpy, matplotlib, opencv, itk, cython,
-joblib, imageio, scikit-image, scikit-learn, imageio-ffmpeg,
-opencv-python>=4
+## Dependencies
+**Required packages:** 
+The complete list of required packages are described in the [requirements.txt](https://github.com/quinngroup/ornet/blob/master/requirements.txt) file.
 
 **Python version:**
-This software works in python versions 3.6 - 3.8
+This software has been tested on python versions 3.6 - 3.8.
 
-# Installation
+## Installation
 Inside of the root directory of this project run:
 
 ```pip install .```
@@ -23,7 +22,7 @@ This installation process should install all of the required dependencies.
 In the event that some of the packages are not installed properly,
 please refer to the dependencies section and manually install all that are listed.
 
-# Testing
+## Testing
 Inside of the "tests" sub-directory, run the following command:
 
 ```
@@ -32,7 +31,7 @@ python ornet_tests.py
 
 7 tests, from various checkpoints along the pipeline, should run without any failures.
 
-# Usage
+## Usage
 Ornet can be utilized by calling the Pipeline module from either the command line interface or in a script.
 Pipeline will output a directory of the following structure:
 
@@ -48,7 +47,7 @@ intermediates contain compressed numpy files (.npz) that store the means, covari
 matrices generated by the gaussian mixture model (GMM) in the pipeline, and the distances directory contains numpy files (.npy)
 that represent the divergence metrics between components of the GMM.
 
-## Command Line Interface:
+### Command Line Interface:
 ```
 python -m ornet.Pipeline -i <input video or directory> -m <mask directory> -o <output directory>
 ```
@@ -59,17 +58,34 @@ For more detailed information regarding command line options the "-h" flag can b
 python -m ornet.Pipeline -h
 ```
 
-## Python script:
+### Python script:
 ```
 from ornet import Pipeline as pipeline
 
 pipeline.run(input_path, mask_path, output_path)
 ```
 
-# Requests, Queries, or Issues
-In case of any requested changes, questions, or issues related to this source code please create an issue here.
+## Community Guidelines
+### Requests or Questions
+For any questions or change requests related to this project, please create an issue [here](https://github.com/quinngroup/ornet/issues). We ask that you use labels, such as "feature request" or "question", when creating an issue.
 
-# Publications
+### Reporting Issues
+If you experinece any unexpected behavior with this software, please document it using this repository's [issues](https://github.com/quinngroup/ornet/issues) section. Please tag issues of this nature with the "unexpected behavior" label.
+
+##### Documentation Process
+1. Post a concise description of the unexpected behavior.
+2. Include a small code snippet that can reproduce the behavior.
+3. Describe the results of the snippet.
+4. Detail the expected results of the snippet.
+
+##### Issue fixes
+If you would like to submit a fix for our developers to review, complete the following steps:
+1. If an issue does not already exist for this problem, create one according to the above documentation process.
+2. Fork the project.
+3. Apply your changes.
+4. Create a pull request, and include in the request a reference to the issue.
+
+## Publications
 If you are using OrNet in a scientific paper, please see the following publications:
 
 [Dynamic Social Network Modeling of Diffuse Subcellular Morphologies](https://conference.scipy.org/proceedings/scipy2018/Andrew_Durden.html)
