### 2020-01-16 97d75dc
simplified command line interface
