+++ setup.py
@@ -6,6 +6,7 @@ setuptools.setup(
 	version='0.1',
 	packages=['ornet', 'ornet.gmm'],
 	package_dir={'ornet':'src', 'ornet.gmm':'src/gmm'},
+	entry_points={'console_scripts': ['ornet = ornet.Pipeline:main']},
 	install_requires=[
 		'itk',
 		'numpy',
