### 2020-01-16 #1983 @serine
## Functionality

- Installation: Does installation proceed as outlined in the documentation?

In the current form software doesn't install. It is missing:

- `opencv-python` module is missing, had to `pip install opencv-python`
I've noticed that `setup.py` has it, but the line is commented out

- `sudo apt-get install python3-tk` This is annoying because it requires system wide installation, could be hard for many users. I recommend looking into packaging with `conda`

- Functionality: Have the functional claims of the software been confirmed?

Work in progress as I haven't been able to get results yet..

- Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

There doesn't appears to be any claims regarding software performance, so this one is ok

## Documentation

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

From reading a README.md in the repo it is hard to get an idea what the tool does. In general I'm still not clear if the tool is for some pre-processing or it is meant to yield some biologically meaningful / interpretable results.

One thing that concerns me is a statement in the paper.md that this is a "prototype" software and there some light hints that a new software might come to exists? I'll try to provide more comments in the next session [Sortware paper](#software-paper) but one of the main purposes of this review is for you to be able to say that you have a functioning software/tool that people in the world can use and we (JOSS reviews) have tested it out and thumbed up. So if you are planing to abundant this tool in the near future, the effort seems wasteful.

- Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

for linux (ubuntu at least) this additional package is required  `python3-tk` system install

It is probably good idea list all python packages under "dependencies" section

- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There is an example folder, however little description given about the video files. It would be nice if there was more explanation about the video files and what the tool will do to the them in terms of processing and the expected results. In addition author seems to assume that the user will know how to do the masking and generate `vtk` file. I think it would be helpful to include a reference link to a external resource on how to generating `vtk` files if one doesn't know. Also some indication of the runtime would be very helpful.

- Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I'm confused about stdout/stderr from the tests. I think those `unclosed file` should be fixed..?

```
2
unclosed file <_io.BufferedWriter name=4>
unclosed file <_io.BufferedReader name=5>
unclosed file <_io.BufferedReader name=7>
...unclosed file <_io.BufferedWriter name=4>
unclosed file <_io.BufferedReader name=5>
unclosed file <_io.BufferedReader name=7>
.unclosed file <_io.BufferedReader name=7>
...
```

- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

## Software paper

- Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

- State of the field: Do the authors describe how this software compares to other commonly-used packages?

- Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

## Other

It would be nice to add detection of the existing output folder or simply warn users that results will be overwritten

There probably need some mention of the resources required for tool to run. i.e 1 cpu or several. Also RAM footprint. I've noticed that on mdivi sample, ornet was taking just over 5Gb. I'm guessing RAM usage will go up with the file (video) size. What is your "typical" amount of raw data and processed? I guess folks in the field should know, however I still feel that this information will be useful. Also mention runtime.. seems like it is taking a long time.
Also is there a need for large tmp/scratch space? Just thinking about running this scrip on the cluster/HPC ..

Does the video format matters? you should probably state file format dependency either way

This is being rather picky, but those spaces in file names are rather annoying for command line tool? Your sample videos have brackets and spaces :).

When you are working with image data, do you get a "movie" file e.g avi, from the instrument or you get bunch of images that you then convert into a "movie" file, which you then use with your tool?
It would be nice to have a more clear view of the process.

I think python version is too stick? The tool seems to work on python version 3.5 and 3.6.7.. It doesn't really matter for the review process, but you sort of limiting amount of people that are potentially could use your tool

Do you need this print statement [src/extract_cells.py:52:    print(number_of_segments)](https://github.com/quinngroup/ornet/blob/e5de22495203d0d24081c47218145fb85cef9f89/src/extract_cells.py#L53) ?

It gives a number on a console without much context.. If this is a meaningful message you should probably prefix a short message i.e
```
print("MSG: Number of segments %s" % number_of_segments, file = sys.stderr)
```
or something similar, or just comment out for debugging purposes later

I think from usability point of view it would be great to change

```
python -m ornet.Pipeline
```

to simply `ornet`. I've send you [PR](https://github.com/quinngroup/ornet/pull/12)
