### 2020-01-16 6e0001b
Added support for coveralls
