+++ .travis.yml
@@ -11,3 +11,5 @@ install:
 script:
   - cd tests
   - coverage run --source=ornet -m unittest discover -s .
+after_success:
+  - coveralls
