### 2020-01-16 856b4c7
Changed the test interface
