### 2020-01-16 543e3db
Changed install test instructions
