### 2020-01-16 #1983 @Marcdh3
Hi @serine. Thank you for the detailed review as well! Many of the issues you have pointed out are actively being fixed as we speak, and we will update everyone once the last fixes are applied. As of now, I believe the last things necessary will be clearing up ambiguity in our documentation and paper, and improving our test suite. A major point needing clarification is that this software will not be abandoned in the near future. We apologize for the confusion caused by the prototype statement. It was intended to indicate that we already have a usable package and we plan on constantly improving many areas of it, but we see how it was misleading.

One interesting issue you discovered was a missing python3-tk package. After receiving feedback from @vc1492a, we've setup continuous integration using travis ci and has been able to successfully test our code against an Ubuntu environment for different python versions using our install instructions. Whenever you find free time, could you please run our latest install instructions without the python3-tk package and let us know if the problem is now resolved.

Once again, we thank everyone for their feedback, because this review feedback has been amazing in terms of improving our work. We are working hard on pushing these changes quickly, so the review process can continue smoothly.
