+++ paper.bib
@@ -28,6 +28,16 @@
     primaryClass={q-bio.SC}
 }
 
+@InProceedings{ Andrew_Durden-proc-scipy-2018,
+  author    = { {A}ndrew {D}urden and {A}llyson {T} {L}oy and {B}arbara {R}eaves and {M}ojtaba {F}azli and {A}bigail {C}ourtney and {F}rederick {D} {Q}uinn and {S} {C}hakra {C}hennubhotla and {S}hannon {P} {Q}uinn },
+  title     = { {D}ynamic {S}ocial {N}etwork {M}odeling of {D}iffuse {S}ubcellular {M}orphologies },
+  booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
+  pages     = { 1 - 7 },
+  year      = { 2018 },
+  editor    = { {F}atih {A}kici and {D}avid {L}ippa and {D}illon {N}iederhut and {M} {P}acer },
+  doi       = { 10.25080/Majora-4af1f417-000 }
+}
+
 @article{Hagberg:2008,
     author={Eric A. Hagberg, Daniel A. Schult and Pieter J. Swart},
     title={Exploring network structure, dynamics, and function using NetworkX},
