### 2020-01-17 #1983 @serine
@Marcdh3 yep, so `python3-tk` isn't required. I've removed that module and installed worked fine. I'm not sure why it isn't required, most likely todo with python version. This is the original issue that I faced

```
(ins)(env) [biostation2]~/.../OrNet/ornet (master)$ python tests/ornet_tests.py
Traceback (most recent call last):
  File "/usr/lib/python3.5/tkinter/__init__.py", line 36, in <module>
    import _tkinter
ImportError: No module named '_tkinter'

Traceback (most recent call last):
  File "tests/ornet_tests.py", line 9, in <module>
    import ornet.Pipeline as pipeline
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/ornet/Pipeline.py", line 19, in <module>
    from ornet.gmm.run_gmm import skl_gmm
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/ornet/gmm/run_gmm.py", line 1, in <module>
    import matplotlib.pyplot as plt
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/matplotlib/pyplot.py", line 2372, in <module>
    switch_backend(rcParams["backend"])
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/matplotlib/pyplot.py", line 207, in switch_backend
    backend_mod = importlib.import_module(backend_name)
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/importlib/__init__.py", line 126, in import_module
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/matplotlib/backends/backend_tkagg.py", line 1, in <module>
    from . import _backend_tk
  File "/home/kirill/projects/JOSS-reviews/OrNet/env/lib/python3.5/site-packages/matplotlib/backends/_backend_tk.py", line 5, in <module>
    import tkinter as Tk
  File "/usr/lib/python3.5/tkinter/__init__.py", line 38, in <module>
    raise ImportError(str(msg) + ', please install the python3-tk package')
ImportError: No module named '_tkinter', please install the python3-tk package
 ```

using python 3.6.7 fixed that issue.

However you should provide more explicit installation steps. This is what I had to do

```
virtualenv -p /opt/python-3.6.7/bin/python3 env
source env/bin/activate
git clone https://github.com/quinngroup/ornet
cd ornet
pip install -r requirements.txt
pip install .
```

```
cd tests
python test_pipeline.py
```

Yep as you've mentioned some work on the docs and you should be all good.

Cheers
