### 2020-01-17 568096a
Chagned to a simpler cli
