### 2020-01-17 42deed4
Added progress bars. Updated bib
