+++ paper.bib
@@ -19,6 +19,14 @@
     doi={10.5281/zenodo.44297}
 }
 
+@software{Costa-Luis:2019,
+    author={Casper da Costa-Luis},
+    title={tqdm: A Fast, Extensible Progress Meter for Python and CLI},
+    journal={Journal of Open Source Software},
+    year={2019},
+    doi={10.21105/joss.01277}
+}
+
 @article{Durden:2019,
     author={Andrew Durden},
     title={Modelling Diffuse Subcellular Protein Structures as Dynamic Social Networks},
