+++ requirements.txt
@@ -24,3 +24,4 @@ scikit-image==0.16.2
 scikit-learn==0.22.1
 scipy==1.4.1
 six==1.13.0
+tqdm==4.41.1
