+++ paper.bib
@@ -20,7 +20,7 @@
 }
 
 @software{Costa-Luis:2019,
-    author={Casper da Costa-Luis},
+    author={{da Costa-Luis}, Casper},
     title={tqdm: A Fast, Extensible Progress Meter for Python and CLI},
     journal={Journal of Open Source Software},
     year={2019},
