+++ src/track_cells.py
@@ -1,0 +1,180 @@
+'''
+This script will take in both a video from a directory and an
+initial masking image of the first frame. It will then create contours for
+each subsequent frame and use those contours to mask out each cell. It will
+then save each cell in their own video
+'''
+# Author : Andrew Durden
+
+import os
+import random
+import argparse
+from functools import partial
+
+import cv2
+import imageio
+import numpy as np
+from tqdm import tqdm
+from matplotlib import pyplot as plt
+
+
+def track_cells(vidfile, maskfile, show_video=False):
+    """
+    reads a video file and initial masks and returns a set of frames for each cell
+
+
+    Parameters
+    ----------
+    vidfile : string
+        path to a single video file
+    maskfile : string
+        path to a single vtk mask file
+    show_video : boolean (Default : False)
+        If true, display video with contours drawn during processing
+
+    Returns
+    ---------
+    videos : Returns a list of arrays each with shape (H, W, F)
+    """
+    im = imageio.imread(maskfile)
+    if show_video:
+        plt.imshow(im)
+        plt.show()
+
+    vf = imageio.get_reader(vidfile)
+    frameNum = 0
+    number_of_segments = len(
+        np.unique(im)) - 1  # defines number of segs from vtk
+    outs = []
+    ims = list()
+    masks = list()
+    colors = list()  # the colors to make the contours in the output video?
+    contours = list()  # holds contours
+    dilates = list()
+    comparisons = list()
+
+    kernel = np.ones((17, 17), np.uint8)  # kernel for opening
+    kernel2 = np.ones((3, 3), np.uint8)  # kernel for dilation
+    font = cv2.FONT_HERSHEY_SIMPLEX  # font for frame count
+        
+    for cols in range(number_of_segments):  # creates random colors to use for the outlines
+        colors.append((random.randint(1, 255), random.randint(0, 255),
+                       random.randint(1, 255)))
+
+    for i in range(number_of_segments):  # separates each mask from the vtk and lists them
+        masks.append(im != i + 1)
+
+    progress_bar = tqdm(total=vf.count_frames())
+    progress_bar.set_description('    Tracking cells')
+    for frameNum, frame in enumerate(vf):  # while( vf.isOpened() ):
+        for i in range(number_of_segments):  # adds a copy of the current frame for each segment
+            ims.append(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY))
+
+        for i in range(number_of_segments):  # blacks out all that isn't in the initial mask for each segment
+            if (frameNum == 0):
+                ims[i][masks[i]] = 0
+            else:
+                ims[i][dilates[
+                           i] != 255] = 0  # uses previous dilation for next frame
+
+        del dilates[:]
+        for i in range(number_of_segments):
+            ret, ims[i] = cv2.threshold(ims[i], 3, 255, cv2.THRESH_BINARY_INV)
+            temp = cv2.morphologyEx(ims[i], cv2.MORPH_OPEN, kernel,
+                                    iterations=3)
+            temp = cv2.bitwise_not(temp)
+            dilates.append(temp)
+
+        # each segment dilates
+        # if two segments overlap cut the overlapping portions from both (first one to cover a place gets it...)
+        if (
+                frameNum % 100 == 1 or frameNum < 20):  # every 100 frames do full pairwise comparisons and save which were adjacent
+            for i in range(5):
+                del comparisons[:]
+                for d in range(number_of_segments):
+                    dilates[d] = cv2.dilate(dilates[d], kernel2, iterations=2)
+                for s in range(
+                        number_of_segments):  # tests each segment with every other segment, slow
+                    newcomps = list()
[…]
