+++ src/extract_cells.py
@@ -1,180 +1,84 @@
 '''
-This script will take in both a video from a directory and an
-initial masking image of the first frame. It will then create contours for
-each subsequent frame and use those contours to mask out each cell. It will
-then save each cell in their own video
+This script serves as a helper to the extract_cells script by taking the 
+segmentation masks generated by it and creating individual videos of
+each cell.
 '''
-# Author : Andrew Durden
+# Author: Marcus Hill
 
 import os
-import random
 import argparse
-from functools import partial
 
 import cv2
 import imageio
 import numpy as np
-from tqdm import tqdm
-from matplotlib import pyplot as plt
 
 
-def extract_cells(vidfile, maskfile, show_video=False):
-    """
-    reads a video file and initial masks and returns a set of frames for each cell
-
+def generate_singles(vid_path, masks_path, output_path, show_vid=False):
+    '''
+    Each individual cell in a video is extracted into it's own video.
 
     Parameters
     ----------
-    vidfile : string
-        path to a single video file
-    maskfile : string
-        path to a single vtk mask file
-    show_video : boolean (Default : False)
-        If true, display video with contours drawn during processing
+    vid_path: String
+        Path to the input video.
+    masks_path: String
+        Path to the segmentation masks for the input video.
+    output_path: String
+        Path to the directory to save the individual videos.
+    show_vid: boolean
+        Flag to show video while extracting cells.
 
     Returns
-    ---------
-    videos : Returns a list of arrays each with shape (H, W, F)
-    """
-    im = imageio.imread(maskfile)
-    if show_video:
-        plt.imshow(im)
-        plt.show()
+    ----------
+    NoneType object
+    '''
 
-    vf = imageio.get_reader(vidfile)
-    frameNum = 0
-    number_of_segments = len(
-        np.unique(im)) - 1  # defines number of segs from vtk
-    outs = []
-    ims = list()
-    masks = list()
-    colors = list()  # the colors to make the contours in the output video?
-    contours = list()  # holds contours
-    dilates = list()
-    comparisons = list()
+    masks = np.load(masks_path)
+    segments = len(np.unique(masks[0])) - 1
 
-    kernel = np.ones((17, 17), np.uint8)  # kernel for opening
-    kernel2 = np.ones((3, 3), np.uint8)  # kernel for dilation
-    font = cv2.FONT_HERSHEY_SIMPLEX  # font for frame count
-        
-    for cols in range(number_of_segments):  # creates random colors to use for the outlines
-        colors.append((random.randint(1, 255), random.randint(0, 255),
-                       random.randint(1, 255)))
+    writers = []
+    vid_name = os.path.split(vid_path)[1].split('.')[0]
+    os.makedirs(output_path, exist_ok=True)
+    reader = imageio.get_reader(vid_path)
 
-    for i in range(number_of_segments):  # separates each mask from the vtk and lists them
-        masks.append(im != i + 1)
+    for i in range(segments):
+        writers.append(cv2.VideoWriter(
+            os.path.join(
+                str(output_path), str(vid_name) + '_' + str(i + 1) + '.avi'),
+            cv2.VideoWriter_fourcc('M', 'J', 'P', 'G'),
+            reader.get_meta_data()['fps'], (masks.shape[1], masks.shape[2])))
 
-    progress_bar = tqdm(total=vf.count_frames())
-    progress_bar.set_description('Tracking cells')
-    for frameNum, frame in enumerate(vf):  # while( vf.isOpened() ):
-        for i in range(number_of_segments):  # adds a copy of the current frame for each segment
-            ims.append(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY))
[…]
