### 2020-01-17 ff9b95a
Constraint is more flexible. More progress bars.
