+++ src/median_normalization.py
@@ -7,6 +7,7 @@ import argparse
 import cv2
 import imageio
 import numpy as np
+from tqdm import tqdm
 
 
 def median_normalize(vid_name, vid_path, out_path):
@@ -26,11 +27,14 @@ def median_normalize(vid_name, vid_path, out_path):
     '''
     medians = []
     reader = imageio.get_reader(vid_path)
+    progress_bar = tqdm(total=(2 * reader.count_frames()))
+    progress_bar.set_description(' Normalizing video')
     for frame in reader:
         grayscale_frame = np.array(cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY))
         flat_frame = grayscale_frame.flatten()
         flat_frame[flat_frame > 0]
         medians.append(np.median(flat_frame))
+        progress_bar.update()
 
     medians = np.array(medians, dtype=np.uint8)
     max_median = np.max(medians)
@@ -43,6 +47,7 @@ def median_normalize(vid_name, vid_path, out_path):
     writer = cv2.VideoWriter(output, 
              cv2.VideoWriter_fourcc('M', 'J', 'P', 'G'),
              fps, size)
+
     for i, frame in enumerate(reader):
         grayscale_frame = np.array(cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY))
         flat_frame = grayscale_frame.flatten()
@@ -50,6 +55,9 @@ def median_normalize(vid_name, vid_path, out_path):
         out_frame = np.array(flat_frame, dtype=np.uint8).reshape(size)
         color_frame = cv2.cvtColor(out_frame, cv2.COLOR_GRAY2RGB)
         writer.write(color_frame)
+        progress_bar.update()
+    
+    progress_bar.close()
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Applies median \
