+++ src/extract_cells.py
@@ -1,5 +1,5 @@
 '''
-This script serves as a helper to the extract_cells script by taking the 
+This script serves as a helper to the track_cells script by taking the 
 segmentation masks generated by it and creating individual videos of
 each cell.
 '''
@@ -13,7 +13,7 @@ import imageio
 import numpy as np
 
 
-def generate_singles(vid_path, masks_path, output_path, show_vid=False):
+def extract_cells(vid_path, masks_path, output_path, show_vid=False):
     '''
     Each individual cell in a video is extracted into it's own video.
 
@@ -81,4 +81,4 @@ if __name__ == "__main__":
     parser.add_argument('-o', '--output', default=os.getcwd(),
                         help="Path to output directory. Default cwd")
     args = vars(parser.parse_args())
-    generate_singles(args['input'], args['masks'], args['output'])
+    extract_cells(args['input'], args['masks'], args['output'])
