### 2020-01-17 fe04d00
Added progress bars for extracting and converting cells
