### 2020-01-17 7a18112
All progress bars are added
