### 2020-01-17 640d227
Updated print statements
