### 2020-01-18 77f603f
Updated tests
