+++ README.md
@@ -35,6 +35,10 @@ python tests.py
 7 tests, from various checkpoints along the pipeline, should run without any failures.
 
 ## Usage
+### Pipeline Outline
+
+<img src="img/ornet_diagram.png" width="1000">
+
 Ornet can be utilized by calling the pipeline module from either the command line interface or in a script.
 pipeline will output a directory of the following structure:
 
