+++ README.md
@@ -37,14 +37,25 @@ python tests.py
 
 7 tests, from various checkpoints along the pipeline, should run without any failures.
 
-## Usage
-### Pipeline Outline
+## Pipeline Outline
 
 <img src="img/pipeline_diagram.png" width="1000">
 
-Ornet can be utilized by calling the pipeline module from either the command line interface or in a script.
-pipeline will output a directory of the following structure:
+**Inputs:**
+OrNet accepts as input 
 
+The pipeline is composed of 7 tasks:
+
+1. Constraining the video (Optional)
+2. Tracking cell movements
+3. Median Normalize each video frame
+4. Downsampling the video (Optional)
+5. Extract individual cells
+6. Computing GMM intermediates
+7. Computing distance metrics
+
+**Outputs**:
+The pipeline will output a directory of the following structure:
 ```
 ├── outputs/
 │   ├── singles/
@@ -54,10 +65,13 @@ pipeline will output a directory of the following structure:
 
 The singles sub-directory will contain the individual videos (.avi) of each extracted cell from the original video, 
 intermediates contain compressed numpy files (.npz) that store the means, covariances, weights, and precisions
-matrices generated by the gaussian mixture model (GMM) in the pipeline, and the distances directory contains numpy files (.npy)
-that represent the divergence metrics between components of the GMM.
+matrices generated by the gaussian mixture model (GMM) in the pipeline, and the distances directory contains numpy files (.npy) that represent the divergence metrics between components of the GMM.
 
-### Command Line Interface:
+## Usage
+
+Ornet can be utilized from either the command line interface or in a script.
+
+**Command Line Interface:**
 ```
 python -m ornet -i <input video or directory> -m <mask directory> -o <output directory> -c <constrain count>
 ```
@@ -68,7 +82,7 @@ For more detailed information regarding the command line options, use the "-h" f
 python -m ornet -h
 ```
 
-### Python script:
+**Python script:**
 ```
 from ornet import pipeline as pipeline
 
@@ -82,7 +96,7 @@ For any questions or change requests related to this project, please create an i
 ### Reporting Issues
 If you experinece any unexpected behavior with this software, please document it using this repository's [issues](https://github.com/quinngroup/ornet/issues) section. Please tag issues of this nature with the "unexpected behavior" label.
 
-***Documentation Process***
+**Documentation Process**
 1. Post a concise description of the unexpected behavior.
 2. Include a small code snippet that can reproduce the behavior.
 3. Describe the results of the snippet.
