### 2020-01-20 9bcbdc5
Updated the pipeline outline section
