### 2020-01-21 9efb1cb
Added more details to describe the inputs
