+++ README.md
@@ -43,7 +43,8 @@ python tests.py
 
 **Inputs:**
 
-The expected input is a single time-series video where the cells are fluorescently tagged to aid in the detection of sub-cellular organisms. (Note: We used a red fluorescent protein, DsRed2-Mito-7, in our experiments.)  
+- A single time-series video where the cells are fluorescently tagged to aid in the detection of sub-cellular organisms. (Note: We used a red fluorescent protein, DsRed2-Mito-7, in our experiments.)  
+- An initial segmentation mask that denotes the location of each cell in the first frame, saved in the (.itk) format. (Documentation for the ITK-Snap tool we used to generate our masks can be found [here](http://www.itksnap.org/pmwiki/pmwiki.php?n=Documentation.SNAP3).)
 
 The **pipeline** is composed of 7 tasks:
 
@@ -67,7 +68,7 @@ The output will be a directory of the following structure:
 
 The singles sub-directory will contain the individual videos (.avi) of each extracted cell from the original video, 
 intermediates contain compressed numpy files (.npz) that store the means, covariances, weights, and precisions
-matrices generated by the gaussian mixture model (GMM) in the pipeline, and the distances directory contains numpy files (.npy) that represent the divergence metrics between components of the GMM.
+matrices generated by the gaussian mixture model (GMM), and the distances directory contains numpy files (.npy) that represent the divergence metrics between components of the GMM.
 
 ## Usage
 
