### 2020-01-21 3a6d8c2
Added details to the pipeline steps
