### 2020-01-21 19b653f
Cleaning up grammar and content
