### 2020-01-21 18755a6
Added ffmpeg env variable via python
