+++ .travis.yml
@@ -4,7 +4,7 @@ python:
   - "3.7"
   - "3.8"
 install:
-  - ffmpeg
+  - sudo apt install -f ffmpeg
   - pip install -U pip
   - pip install -r requirements.txt   
   - pip install coverage coveralls
