### 2020-01-22 6919337
Removed ffmpeg fixes and pip upgrades
