+++ .travis.yml
@@ -4,8 +4,6 @@ python:
   - "3.7"
   - "3.8"
 install:
-  - sudo apt install -f ffmpeg
-  - pip install -U pip
   - pip install -r requirements.txt   
   - pip install coverage coveralls
   - pip install .
