+++ .travis.yml
@@ -4,6 +4,7 @@ python:
   - "3.7"
   - "3.8"
 install:
+  - pip install -U pip
   - pip install -r requirements.txt   
   - pip install coverage coveralls
   - pip install .
