+++ requirements.txt
@@ -17,6 +17,7 @@ networkx==2.4
 numpy==1.18.1
 opencv-python==4.1.2.30
 Pillow==7.0.0
+pip==19.3
 pyparsing==2.4.6
 python-dateutil==2.8.1
 PyWavelets==1.1.1
