### 2020-01-22 c74220d
Renamed samples. Added a downsample cli option
