### 2020-01-22 165c43a
Updated with a downsample option
