+++ README.md
@@ -94,6 +94,8 @@ The singles sub-directory will contain the individual videos (.avi) of each extr
 intermediates contain compressed numpy files (.npz) that store the means, covariances, weights, and precisions
 matrices generated by the gaussian mixture model (GMM), and the distances directory contains numpy files (.npy) that represent the divergence metrics between components of the GMM.
 
+Note: The pipeline generates temporary files that are deleted upon completion of all tasks. In our experiments, we noticed that our videos comprised of around 20,000 frames used approximately 3.5GB of disk space for the temporary files. The final output directory size was approximately 4MB. Ultimately, we expect the amount temporary space needed will grow proportionally with the size of input video.  
+
 ## Usage
 
 Ornet can be utilized from either the command line interface or in a script.
