### 2020-01-29 24ccc2a
Update paper.md
