+++ paper.md
@@ -34,13 +34,14 @@ bibliography: paper.bib
 
 # Summary
 
-Mitochondria can offer key insights into the health of a human cell. These organelles are strung out across a cell in diffuse, fluctuating webs; healthy structures can be altered by disease, cellular invaders, and other malfunctioning organelles. Fluorescent tagging has shed light on mitochondria’s spatial orientation, but practical observations of structural changes are limited. Biomedical and clinical researchers cannot take advantage of large datasets (i.e. videos of mitochondria in various environments), without automated methods to quantify distribution changes over time. Fitting a model to individual video frames and comparing how the scene evolves has shown some success in measuring spatiotemporal changes [@Ruan:2019]. 
+Mitochondria can offer key insights into the health of a human cell. These organelles are strung out across a cell in diffuse, fluctuating webs; healthy structures can be altered by disease, cellular invaders, and other malfunctioning organelles [@Stavru:2016]. Previous work using mitochondrial distribution focuses on building a model of the cell’s morphology. For instance, both the Allen Cell Structure Segmenter [@Chen:2016] and CellOrganizer [@Murphy:2015] toolkits generate models for organelle structure, providing visual or probabilistic representations. Successful generation of static models has led to investigations in how structures develop over time, such as using mitochondrial distributions to map cells as they grow into a specialized form [@Ruan:2019].  This describes a single, albeit complex cellular process; many more processes could be described by mitochondrial distributions. 
 
-Our project builds on this idea [@Durden:2019], as we seek to convert physical mitochondrial networks into graph networks, then characterize distinct phenotypes. This tool kit is a prototype for future analysis kits. Specifically, it focuses on characterizing two known mitochondrial conditions: drastic fragmentation and excessive fusion. This may assist researchers in larger clinical studies, but expanding this project might eventually yield tools to pick up on structural changes that elude the human eye.
+Our project builds on this idea [@Durden:2019], as we seek to convert physical mitochondrial networks into graph networks, then characterize distinct phenotypes. This tool kit is a prototype for future analysis kits. As it stands, we focus on characterizing two known mitochondrial conditions: drastic fragmentation and excessive fusion. Our resulting model seeks to pull out the key features that make up these processes from a data set, rather than immediately classify the set. This creates a staging ground for biologists and clinicians to use our pipeline to examine evolving mitochondrial distributions in more detail.
 
-OrNet, a python package [@Python:1995], relies heavily on algorithms from the Scikit-Learn tool kit [@scikit-learn:2011]. This in turn is built on NumPy [@Numpy:2006], SciPy [@2019arXiv190710121V], and Matplotlib [@Hunter:2007]. Our program takes videos of human cells with fluorescently tagged mitochondria and first extracts individual cells. With that base to work from, a model is fitted to video frames by calculating a mixture of probability masses (Gaussian Mixture Model). Essentially, this model identifies the center of the densest regions of mitochondrial protein, or peaks, and finds the probability that the surrounding protein belongs to a specific local distribution. The ultimate goal is to analyze the data as a graph, where the peaks in density are used as nodes, and probability metrics form weights and edges between nodes. 
[…]
