+++ paper.bib
@@ -192,3 +192,38 @@ archivePrefix = {arXiv},
     year={2002},
     publisher={IOS Press Amsterdam}
 }
+
+@article {Chen:2016,
+	author = {Chen, Jianxu and Ding, Liya and Viana, Matheus P. and Hendershott, Melissa C. and Yang, Ruian and Mueller, Irina A. and Rafelski, Susanne M.},
+	title = {The Allen Cell Structure Segmenter: a new open source toolkit for segmenting 3D intracellular structures in fluorescence microscopy images},
+	elocation-id = {491035},
+	year = {2018},
+	doi = {10.1101/491035},
+	publisher = {Cold Spring Harbor Laboratory},
+	URL = {https://www.biorxiv.org/content/early/2018/12/08/491035},
+	eprint = {https://www.biorxiv.org/content/early/2018/12/08/491035.full.pdf},
+	journal = {bioRxiv}
+}
+
+
+@article {Stavru:2011,
+    author = {Stavru, Fabrizia and Cossart, Pascale},
+    year = {2011},
+    month = {05},
+    pages = {364-6},
+    title = {Listeria infection modulates mitochondrial dynamics},
+    volume = {4},
+    journal = {Communicative & integrative biology},
+    doi = {10.4161/cib.4.2.15506}
+}
+
+@article{Murphy:2015,
+    author = {Murphy, Robert},
+    year = {2015},
+    month = {10},
+    pages = {},
+    title = {Building Cell Models and Simulations from Microscope Images},
+    volume = {96},
+    journal = {Methods (San Diego, Calif.)},
+    doi = {10.1016/j.ymeth.2015.10.011}
+}
