+++ README.md
@@ -1,5 +1,6 @@
 # OrNet [![Build Status](https://travis-ci.org/quinngroup/ornet.svg?branch=master)](https://travis-ci.org/quinngroup/ornet) ![Python status](https://img.shields.io/badge/python-3.6%20%7C%203.7%20%7C%203.8-blue) [![status](https://joss.theoj.org/papers/6fbce182256576a0dd3d35c26f83a9ef/status.svg)](https://joss.theoj.org/papers/6fbce182256576a0dd3d35c26f83a9ef)
-**OrNet** is a python pipeline for analyzing fluorescence microscopy imagery of organelles. 
+**OrNet** is a python pipeline for analyzing 
+microscopy imagery of organelles. 
 
 
 <img src="img/MitoEx.png" width="400">
@@ -7,7 +8,7 @@
 *Mitochondria in a single cell, with red flourescent tagging* 
 
 ## Purpose
-Analyzing the morphological changes organelles undergo after exposure to external stimul is significant because some structures offer crucial insight into the manner in which human cells responds to viral or bacterial infections. This framework is intended to mitigate the need for manual inspection of flouresence microscopy videos to identify time points and frame regions where significant changes are occurring.
+Analyzing the morphological changes organelles undergo after exposure to external stimuli is significant because some structures offer crucial insight into the manner in which human cells responds to viral or bacterial infections. This framework is intended to mitigate the need for manual inspection of fluorescence microscopy videos to identify time points and frame regions where significant changes are occurring.
 
 ## Dependencies
 **Required packages:** 
@@ -43,7 +44,7 @@ python tests.py
 
 **Concept:**
 
-Inspired by social networks, OrNet models the dynamic morphologies of organelles as social network graphs. Just as Twitter and Facebook capture the dynamic nature of interpersonal relationships through the establishment or dissolution of connections between individuals, OrNet creates graphs that capture the spatio-temporal relationships of organellar structures via a similar edge management process. Graphs are constructed by fitting gaussian mixture models to every frame in a video; the final means become the vertices, while a divergenece metric is applied to every combination pair of mixture components to create the edges. Once graphs are created for each frame, spectral decompostion can be utilized to track the leading eigenvalues to understand the time-points and frame regions where the organellar structures are demonstrating significant changes.      
+Inspired by social networks, OrNet models the dynamic morphologies of organelles as social network graphs. Just as Twitter and Facebook capture the dynamic nature of interpersonal relationships through the establishment or dissolution of connections between individuals, OrNet creates graphs that capture the spatio-temporal relationships of organellar structures via a similar edge management process. Graphs are constructed by fitting gaussian mixture models to every frame in a video; the final means become the vertices, while a divergence metric is applied to every combination pair of mixture components to create the edges. Once graphs are created for each frame, spectral decomposition can be utilized to track the leading eigenvalues to understand the time-points and frame regions where the organellar structures are demonstrating significant changes.      
 
 **Inputs:**
 
@@ -66,15 +67,15 @@ The **pipeline** is composed of 7 tasks:
 
 4. Downsample the video and masks (Optional)
 
-    Skip a given number of frames in the both the video and masks to generate a smaller video. This is useful for vidoes         where cells slowly move over time, thus not any significant change is detected between many of the frames. 
+    Skip a given number of frames in the both the video and masks to generate a smaller video. This is useful for videos         where cells slowly move over time, thus not any significant change is detected between many of the frames. 
 
 5. Extract individual cells
 
[…]
