### 2020-02-12 96c31fa
Added the eigenspectrum script
