+++ src/eigenspectrum.py
@@ -1,0 +1,198 @@
+'''
+This script creates vizualizations of the eigenspectrum 
+data for analysis.
+'''
+#Author: Marcus Hill
+
+import os
+import sys
+import csv
+import argparse
+
+from tqdm import tqdm
+import numpy as np
+import scipy.linalg as sla
+import matplotlib.pyplot as plt
+from scipy.sparse import csgraph
+
+def compute_similarity(matrix, beta=5):
+    '''
+    Applies a heat kernel to a divergence matrix to convert
+    it to a similarity/affinity matrix.
+
+    Parameters
+    ----------
+    matrix: numpy array of floats
+        divergence matrix to be computed into a similarity matrix.
+
+    Returns
+    ----------
+    similarity: numpy array of floats
+        similarity matrix of the input.
+    '''
+    return np.exp((-beta * matrix) / np.std(matrix))
+
+def eigen_decomposition(matrix):
+    '''
+    Perform eigendecomposition on the input graph matrix.
+
+    Parameters
+    ----------
+    matrix: 2-d numpy array
+        Weighted graph adjacency matrix.
+
+    Returns
+    ---------
+    eigen_vals: numpy array
+        Vector of the graph laplacian eigenvalues.
+        Sorted from largest to smallest.
+    eigen_vecs: 2-d numpy array
+        Matrix of the graph laplacian eigenvectors.
+        Sorted from largest to smallest.
+    '''
+
+    affinity = compute_similarity(matrix)
+    laplacian = csgraph.laplacian(affinity)
+    eigen_vals, eigen_vecs = sla.eigh(laplacian)
+
+    sorted_indices = list(range(len(eigen_vals)).__reversed__())
+    eigen_vals = eigen_vals[sorted_indices]
+    eigen_vecs = eigen_vecs[:, sorted_indices]  
+
+    return eigen_vals, eigen_vecs
+
+def eigenspectrum_plot(args):
+    '''
+    Plots the eigenspectrum data across all video frames.
+
+    Parameters
+    ----------
+    args: dict
+        Parsed cli arguments. Details the input video(s)
+        path(s) and output directory.
+
+    Returns
+    ----------
+    NoneType object
+    '''
+
+    progress_bar = tqdm(total=len(args['input']))
+    for vid in args['input']:
+        frames = np.load(vid)
+        vid_eigenvals = []
+        for graph_matrix in frames:
+            eigen_vals, eigen_vecs = eigen_decomposition(graph_matrix)
+            vid_eigenvals.append(eigen_vals)
+
+        vid_name = os.path.split(vid)[-1].split('.')[0]
+        vid_eigenvals = np.array(vid_eigenvals)
+        plt.plot(vid_eigenvals)
+        plt.suptitle(vid_name)
+        plt.savefig(os.path.join(args['output'], vid_name))
+        progress_bar.update()
+
+    progress_bar.close()
+
+def generate_dataset(args):
+    '''
+    Accepts a directory of cell distance files and creates a 
+    supervised learning dataset to classify each cell type 
+    by eigenspectrum features. 
+    '''
+
+    output = os.path.join(args['output'], 'dataset.csv')
+    with open(output, 'w') as fp:
+        writer = csv.writer(fp)
+        writer.writerow(('mean_diff', 'mean_std', 'mean_abs_change', 
+                         'mean_condition_number', 'class'))
+        for vid_name in os.listdir(args['input']):
+            vid = np.load(os.path.join(args['input'], vid_name))
+            affinities = [compute_similarity(x) for x in vid]
+            laplacians = [csgraph.laplacian(x) for x in affinities]
+            eigens = [sla.eigh(x) for x in laplacians]
+            eigvals = []
[…]
