### 2020-02-13 87c0dcb
Modified the code to accept directories of graph matrices.
