+++ paper.bib
@@ -227,3 +227,10 @@ archivePrefix = {arXiv},
     journal = {Methods (San Diego, Calif.)},
     doi = {10.1016/j.ymeth.2015.10.011}
 }
+
+@article{Eliceiri:2012,
+    author={Eliceiri KW, Berthold MR, Goldberg IG, Ibáñez L, Manjunath BS, Martone ME, Murphy RF, Peng H, Plant AL, Roysam B, Stuurman N, Swedlow JR, Tomanack P, Carpenter AE},
+    title={Biological imaging software tools},
+    journal={Nature Methods},
+    year={2012}
+}
