### 2020-02-13 a03a47b
Updating the gitignore
