### 2020-02-13 43ce61e
Removed unecessary files
