### 2020-02-13 #1983 @Marcdh3
The paper and repo have been updated to reflect all of the suggestions from everyone. Thanks for all of the patience. Please let us know if there is anything else that still needs fixing.
