### 2020-02-15 #1983 @vc1492a
Thanks @Marcdh3! I checked out the updated repository, the updates look great to me - thanks for integrating those. I do not see any other open issues from the notes above and the code and documentation is much improved.

I commented on one or two of the open issues as I believe they can now be closed. In addition, I think it would be great if you all could host the package on the Python Package Index in the near future. This would further open up the software to the community and ease the download process as well, especially when adding ornet to a pipeline. Type hinting may also be something you could add in the future - just throwing out some ideas. I don't think these are required for publication in JOSS (and thus we can proceed forward from my perspective), but I would highly encourage publishing to the package index and strictly adhering to a versioning scheme when pushing updates to the master branch.

This is great and interesting work, cheers!
