### 2020-02-18 482172b
Saves the rgb single videos, instead of grayscale
