### 2020-02-19 4f542f1
Updated paper.bib
