+++ paper.bib
@@ -98,8 +98,7 @@ title = {Getting Started with Pyparsing},
 year = {2007},
 isbn = {9780596514235},
 publisher = {O’Reilly},
-edition = {First},
-doi={10.5555/1406599}
+edition = {First}
 }
 
 @Misc{Numpy:2006,
@@ -213,8 +212,7 @@ archivePrefix = {arXiv},
     pages = {364-6},
     title = {Listeria infection modulates mitochondrial dynamics},
     volume = {4},
-    journal = {Communicative & integrative biology},
-    doi = {10.4161/cib.4.2.15506}
+    journal = {Communicative & integrative biology}
 }
 
 @article{Murphy:2015,
