+++ paper.bib
@@ -212,7 +212,8 @@ archivePrefix = {arXiv},
     pages = {364-6},
     title = {Listeria infection modulates mitochondrial dynamics},
     volume = {4},
-    journal = {Communicative & integrative biology}
+    journal = {Communicative & integrative biology},
+    doi = {10.4161/cib.4.3.15506}
 }
 
 @article{Murphy:2015,
