+++ src/eigenspectrum.py
@@ -5,12 +5,15 @@ data for analysis.
 #Author: Marcus Hill
 
 import os
+import re
 import sys
 import csv
+import math
 import argparse
 
-from tqdm import tqdm
+import imageio
 import numpy as np
+from tqdm import tqdm
 import scipy.linalg as sla
 import matplotlib.pyplot as plt
 from scipy.sparse import csgraph
@@ -31,6 +34,15 @@ def compute_similarity(matrix, beta=5):
         similarity matrix of the input.
     '''
     return np.exp((-beta * matrix) / np.std(matrix))
+
+def sort_eigens(eigen_vals, eiegen_vecs):
+    '''
+    '''
+    sorted_indices = np.argsort(eigen_vals)
+    eigen_vals = eigen_vals[sorted_indices]
+    eigen_vecs = eigen_vecs[:, sorted_indices]
+
+    return eigen_vals, eigen_vecs
 
 def eigen_decomposition(matrix):
     '''
@@ -53,12 +65,10 @@ def eigen_decomposition(matrix):
 
     affinity = compute_similarity(matrix)
     laplacian = csgraph.laplacian(affinity)
-    eigen_vals, eigen_vecs = sla.eigh(laplacian)
+    #eigen_vals, eigen_vecs = sla.eigh(laplacian)
+    #sorted_indices = list(range(len(eigen_vals)).__reversed__())
 
-    sorted_indices = list(range(len(eigen_vals)).__reversed__())
-    eigen_vals = eigen_vals[sorted_indices]
-    eigen_vecs = eigen_vecs[:, sorted_indices]  
-
+    eigen_vals, eigen_vecs = np.linalg.eig(laplacian)
     return eigen_vals, eigen_vecs
 
 def eigenspectrum_plot(vids, outdir, max_vals=10):
@@ -85,6 +95,7 @@ def eigenspectrum_plot(vids, outdir, max_vals=10):
         vid_eigenvals = []
         for graph_matrix in frames:
             eigen_vals, eigen_vecs = eigen_decomposition(graph_matrix)
+            eigen_vals, eigen_vecs = sort_eigens(eigen_vals, eigen_vecs)
             vid_eigenvals.append(eigen_vals[:max_vals])
 
         plt.suptitle(vid_name)
@@ -93,6 +104,40 @@ def eigenspectrum_plot(vids, outdir, max_vals=10):
         progress_bar.update()
 
     progress_bar.close()
+
+def regions_of_interest(distances, intermediates, videos, outdir, std_count=3):
+    '''
+    '''
+    for distance_file_path in distances:
+        filename = os.path.split(distance_file_path)[-1].split('.')[0]
+
+        for inter_file_path in intermediates:
+            if filename in inter_file_path:
+                inter = np.load(inter_file_path)
+
+        for vid_file_path in videos:
+            video_name  = re.sub('_gray', '', filename)
+            if video_name in vid_file_path:
+                vid = list(imageio.get_reader(vid_file_path))
+        
+        #Write a for loop iterating through
+        #graph_matrices = np.load(distance_file_path)
+        means, covars = inter['means'], inter['covars']
+        color = [np.random.randint(256), np.random.randint(256), np.random.randint(256)]
+        #for frame in range(len(means)):
+        for frame in range(1):
+            for j, mean in enumerate(means[frame]):
+                x_diff = std_count * math.sqrt(covars[frame][j][0][0])
+                y_diff = std_count * math.sqrt(covars[frame][j][1][1])
+                x_bounds = [int(mean[0] - x_diff), int(mean[0] + x_diff)]
[…]
