### 2020-02-21 d42c1e2
Refactored eigenspectrum into an anaylsis package.
