### 2020-02-21 2db07f4
Writes output videos
