+++ src/analysis/tracking.py
@@ -10,14 +10,34 @@ import csv
 import math
 import argparse
 
+import cv2
 import imageio
 import numpy as np
+from tqdm import tqdm
 import matplotlib.pyplot as plt
+
+from .util import spectral_decomposition, sort_eigens
 
 def regions_of_interest(distances, intermediates, videos, outdir, std_count=3):
     '''
+    Draws bounding boxes around frame regions demonstrating
+    the highest amounts of variance.
+
+    Parameters
+    ----------
+    distances: list
+        Paths to weighted adjacency matrices (.npy).
+    intermediates: list
+        Paths to gmm intermediates (.npz).
+    videos: list
+        Paths to single cell videos (.avi).
+    outdir: string
+        Path to directory to save output video.
+    std_count: int
+        Number of standard deviations away from the means
+        to draw the bounding box borders.
     '''
-    for distance_file_path in distances:
+    for distance_file_path in tqdm(distances):
         filename = os.path.split(distance_file_path)[-1].split('.')[0]
 
         for inter_file_path in intermediates:
@@ -27,35 +47,52 @@ def regions_of_interest(distances, intermediates, videos, outdir, std_count=3):
         for vid_file_path in videos:
             video_name  = re.sub('_gray', '', filename)
             if video_name in vid_file_path:
-                vid = list(imageio.get_reader(vid_file_path))
+                reader = imageio.get_reader(vid_file_path)
+                fps = reader.get_meta_data()['fps']
+                size = reader.get_meta_data()['size']
+                vid = list(reader)
         
-        '''
         graph_matrices = np.load(distance_file_path)
         for i, matrix in enumerate(graph_matrices):
             eigen_vals, eigen_vecs = spectral_decomposition(matrix)
             eigen_vals, eigen_vecs, sorted_indices = sort_eigens(
                 eigen_vals, eigen_vecs)
-            # Line 133-134: Remove frame loop
-            # Line 135: for j in sorted_indices
-        '''
-        means, covars = inter['means'], inter['covars']
-        color = [np.random.randint(256), np.random.randint(256), np.random.randint(256)]
-        #for frame in range(len(means)):
-        for frame in range(1):
-            for j, mean in enumerate(means[frame]):
-                x_diff = std_count * math.sqrt(covars[frame][j][0][0])
-                y_diff = std_count * math.sqrt(covars[frame][j][1][1])
-                x_bounds = [int(mean[0] - x_diff), int(mean[0] + x_diff)] #means[i][j][0]
-                y_bounds = [int(mean[1] - y_diff), int(mean[1] + y_diff)] #means[i][j][1]
-            
-            vid[frame][x_bounds[0]:x_bounds[1], y_bounds[0], :] = color
-            vid[frame][x_bounds[0]:x_bounds[1], y_bounds[1], :] = color
-            vid[frame][x_bounds[0], y_bounds[0]:y_bounds[1], :] = color
-            vid[frame][x_bounds[1], y_bounds[0]:y_bounds[1], :] = color
-            plt.imshow(vid[frame])
-            plt.show()
+            means, covars = inter['means'], inter['covars']
+            color = [np.random.randint(256), np.random.randint(256), np.random.randint(256)]
+            for j in sorted_indices[:1]:
+                x_diff = std_count * math.sqrt(covars[i][j][0][0])
+                y_diff = std_count * math.sqrt(covars[i][j][1][1])
+                x_bounds = [int(means[i][j][0] - x_diff), int(means[i][j][0] + x_diff)]
+                y_bounds = [int(means[i][j][1] - y_diff), int(means[i][j][1] + y_diff)]
 
+                if x_bounds[0] < 0:
+                    x_bounds[0] = 0
 
[…]
