### 2020-02-21 802d691
Added FPS options
