+++ paper.md
@@ -38,9 +38,9 @@ bibliography: paper.bib
 
 # Summary
 
-Fluorescent microscopy videos are vital for analyzing the morphological changes that subcellular protein structures undergo after exposure to external stimuli. Changes in organelle structures offer crucial insight into the manner in which cells respond to viral or bacterial infections, cellular invaders, or even the organelles themselves malfunctioning [@Stavru:2011]. Generally, modeling organellar structures involve manually inspecting each video then denoting time points and regions that demonstrate anomalous behavior. However, manual techniques are error prone and not scalable to high volumes of data. Thus, arises the need to find a methodology that generates quantitative models capable of accurately describing the data [@Eliceiri:2012]. Prior works have demonstrated successful generation of static models [@Murphy:2015; @Chen:2016; @Ruan:2019]. Such advancements have inspired us to propose a novel framework, OrNet, that models both the spatial and temporal morphology changes that organelles undergo as dynamic social networks.
+Fluorescent microscopy videos are vital for analyzing the morphological changes that subcellular protein structures undergo after exposure to external stimuli. Changes in organelle structures offer crucial insight into the manner in which cells respond to viral or bacterial infections, cellular invaders, or even the organelles themselves malfunctioning [@Stavru:2011]. Generally, modeling organellar structures involve manually inspecting each video then denoting time points and regions that demonstrate anomalous behavior. However, manual analyses lack objective metrics to assess morphological changes, and thus hinder the ability to perform secondary analyses and quantitative comparisons. Thus, arises the need to find a methodology that generates quantitative models capable of accurately describing the data [@Eliceiri:2012]. Prior works have demonstrated success in the generation of static models for subcellular modeling [@Murphy:2015; @Chen:2016; @Ruan:2019]. Such advancements have inspired us to propose a novel framework, OrNet, that models both the spatial and temporal morphology changes that organelles undergo as dynamic social networks.
 
-OrNet is an open-source python package [@Python:1995] that is built-upon the libraries of Scikit-Learn [@scikit-learn:2011], NumPy [@Numpy:2006], SciPy [@2019arXiv190710121V], and Matplotlib [@Hunter:2007]. Our tool accepts as input microscopy videos of cells with fluorescently tagged organelles, and outputs quantitative descriptions of the morphological changes. The methodology consists of developing social networks that capture the spatio-temporal relationships of organellar structures via a dynamic edge management process. The graphs are constructed by fitting gaussian mixture models to every frame of an input video; the final means become the vertices, while a divergence metric is applied to every combination pair of mixture components to create the edges. Once graphs are created for each frame, spectral decomposition is utilized to track the leading eigenvalues to understand the time-points and frame regions where organellar structures are demonstrating significant changes. 
[…]
