+++ src/extract_cells.py
@@ -3,7 +3,6 @@ This script serves as a helper to the track_cells script by taking the
 segmentation masks generated by it and creating individual videos of
 each cell.
 '''
-# Author: Marcus Hill
 
 import os
 import argparse
