+++ src/__main__.py
@@ -25,7 +25,7 @@ def parse_cli(args):
     '''
 
     parser = argparse.ArgumentParser(
-        usage = 'python -m ornet [-h] -i INPUT -m MASKS -o OUTPUT',
+        usage = 'python -m ornet [-h] -i INPUT -m MASKS -o OUTPUT -c CONSTRAIN -d DOWNSAMPLE',
         description='An end-to-end pipeline of OrNet.'
     )
     parser.add_argument('-i', '--input',
@@ -36,7 +36,7 @@ def parse_cli(args):
                         required=True)
     parser.add_argument('-o', '--output', default=os.getcwd(),
                         help='Output directory to save files.')
-    parser.add_argument('-c', '--count', type=int, default=-1,
+    parser.add_argument('-c', '--constrain', type=int, default=-1,
                         help='First N frames of video to use. Default is all.')
     parser.add_argument('-d', '--downsample', type=int, default=1,
                         help='The number of frames to skip when performing'
@@ -46,7 +46,7 @@ def parse_cli(args):
 
 def main(system_args):
     args = parse_cli(system_args[1:])
-    pipeline.run(args['input'], args['masks'], args['output'], args['count'], 
+    pipeline.run(args['input'], args['masks'], args['output'], args['constrain'], 
                  args['downsample'])
 
 if __name__ == '__main__':
