+++ src/analysis/classification.py
@@ -1,0 +1,165 @@
+'''
+Generates a supervsied learning dataset of the
+eigenspectrum data using fourier transforms.
+'''
+
+import os
+import sys
+import csv
+import argparse
+
+import numpy as np
+from tqdm import tqdm
+import matplotlib.pyplot as plt
+from scipy.fftpack import fft, fftfreq, ifft
+
+from ornet.analysis.util import sort_eigens
+
+def generate_frequency_dataset(eigen_dir_path, outfile_path, 
+                               label, plot_dir=None, k=None):
+    '''
+    Applies to Fourier Transforms to the leading
+    eigenvalues, and uses the frequencies as features
+    to construct a supervised learning dataset.
+
+    Parameters
+    ----------
+    eigen_dir_path: string
+        Path to the directory containing eigen data (.npz).
+    outfile_path: string
+        Path to save the dataset (.csv).
+    label: string
+        Class label for the instances.
+    plot_dir: string
+        If a path is given the frequency plots
+        corresponding to each eigenvalue time series
+        is saved at the given directory path.
+    k: int
+        First k leading eigenvalues. Default is
+        all eigenvalues.
+    
+    Returns
+    -------
+    NoneType object
+    '''
+
+    with open(outfile_path, 'a+') as fp:
+        writer = csv.writer(fp)
+        headers = ['frequency_' + str(x) for x in np.arange(100)] 
+        headers.append('class')
+        writer.writerow(headers)
+
+        for file_path in tqdm(os.listdir(eigen_dir_path)):
+            file_name = file_path.split('.')[0]
+            data = np.load(os.path.join(eigen_dir_path, file_path))
+            eigen_vals, eigen_vecs = data['eigen_vals'], data['eigen_vecs']
+
+            if(eigen_vals.shape[0] == 200):
+                sorted_eigen_vals = np.empty(eigen_vals.shape)
+                for i in range(len(eigen_vals)):
+                    current_eigen_vals, current_eigen_vecs, sorted_indices = sort_eigens(
+                        eigen_vals[i], 
+                        eigen_vecs[i]
+                    )
+                    sorted_eigen_vals[i] = current_eigen_vals
+
+                if k == None or k > sorted_eigen_vals.shape[1]:
+                    max_eigenvals = sorted_eigen_vals.shape[1]
+                else:
+                    max_eigenvals = k
+
+                sampling_rate = len(sorted_eigen_vals)
+                first_half = len(sorted_eigen_vals) // 2
+                magnitude_sum = np.zeros(first_half)
+
+                for i in range(max_eigenvals):
+                    #Compute the frequencies and store the first half
+                    eigen_fft = fft(sorted_eigen_vals[:, i])
+                    freqs = fftfreq(len(sorted_eigen_vals)) * sampling_rate
+                    freqs = freqs[:first_half] 
+
+                    #Compute the magnitude of the singal and normalize
+                    magnitude = np.abs(eigen_fft) * (1/len(eigen_fft)) 
+                    magnitude = magnitude[:first_half]
+                    magnitude_sum += magnitude
+
+                    #Write to CSV
+                    magnitude_list = list(magnitude_sum)
+                    magnitude_list.append(label)
+                    writer.writerow(magnitude_list)
+
+                    #(Optional) Save the eigenvalue frequency plots
+                    if plot_dir != None:
+                        plot_title = file_name + '_leading_eigenval_' + str(i)
+                        plt.figure()
+                        plt.title(plot_title)
+                        plt.bar(freqs, magnitude, width=1.5)
+                        plt.savefig(os.path.join(plot_path, plot_title + '.png'))
+                        plt.close()
+
+def parse_cli(args):
+    '''
+    Parses command line arguments.
+    
+    Parameters
+    ----------
+    args: list
+        Arguments from the command line.
+
[…]
