### 2020-03-17 25b73dd
Moved csv writer operation outside of loop.
