### 2020-03-18 #1983 @Marcdh3
@arfon @akeshavan I understand the submission process is undergoing changes in light of the pandemic, so I would like to inquire the status of our submission. I believe the process to being accepted was initiated approximately two weeks ago, but Whedon has not taken any actions.
